[refactor] DTO 디렉토리 위치 변경
api -> api/family 도메인으로 변경
