+++ src/main/java/com/suitecare/suitecare/api/family/controller/FamilyController.java
@@ -1,8 +1,8 @@
 package com.suitecare.suitecare.api.family.controller;
 
 import com.suitecare.suitecare.api.family.service.FamilyService;
-import com.suitecare.suitecare.domain.CreateFamilyDTO;
-import com.suitecare.suitecare.domain.LoginDTO;
+import com.suitecare.suitecare.api.family.domain.CreateFamilyDTO;
+import com.suitecare.suitecare.api.family.domain.LoginDTO;
 import jakarta.servlet.http.HttpServletResponse;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
