+++ src/main/java/com/suitecare/suitecare/api/family/repository/FamilyMapper.java
@@ -1,8 +1,8 @@
 package com.suitecare.suitecare.api.family.repository;
 
-import com.suitecare.suitecare.domain.CreateFamilyDTO;
-import com.suitecare.suitecare.domain.FamilyDTO;
-import com.suitecare.suitecare.domain.LoginDTO;
+import com.suitecare.suitecare.api.family.domain.CreateFamilyDTO;
+import com.suitecare.suitecare.api.family.domain.FamilyDTO;
+import com.suitecare.suitecare.api.family.domain.LoginDTO;
 import org.apache.ibatis.annotations.Mapper;
 
 @Mapper
